PROJECT DEVELOPMENT PHASE/SPRINT 3/requirements.txt
asttokens==2.1.0
backcall==0.2.0
backports.functools-lru-cache==1.6.4
bcrypt==4.0.1
certifi==2022.9.24
click==8.1.3
colorama==0.4.6
contourpy==1.0.6
cycler==0.11.0
debugpy==1.6.3
decorator==5.1.1
entrypoints==0.4
executing==1.2.0
Flask==2.2.2
Flask-Bcrypt==1.0.1
Flask-Cors==3.0.10
Flask-Login==0.6.2
Flask-MySQLdb==1.0.1
Flask-WTF==1.0.1
fonttools==4.38.0
ipykernel==6.17.0
ipython==8.6.0
itsdangerous==2.1.2
jedi==0.18.1
Jinja2==3.1.2
joblib==1.2.0
jupyter_client==7.4.4
jupyter_core==4.11.2
kiwisolver==1.4.4
MarkupSafe==2.1.1
matplotlib==3.6.2
matplotlib-inline==0.1.6
mysql-connector==2.2.9
mysql-connector-python==8.0.31
mysqlclient==2.1.1
nest-asyncio==1.5.6
numpy==1.23.4
packaging==21.3
pandas==1.5.1
parso==0.8.3
pickleshare==0.7.5
Pillow==9.3.0
pip==21.2.4
prompt-toolkit==3.0.32
protobuf==3.20.1
psutil==5.9.0
pure-eval==0.2.2
Pygments==2.13.0
pyparsing==3.0.9
python-dateutil==2.8.2
pytz==2022.5
pywin32==302
pyzmq==23.2.0
scikit-learn==1.1.2
scipy==1.9.3
seaborn==0.12.1
setuptools==61.2.0
six==1.16.0
sklearn==0.0
stack-data==0.6.0
threadpoolctl==3.1.0
tornado==6.2
traitlets==5.5.0
wcwidth==0.2.5
Werkzeug==2.2.2
wheel==0.37.1
wincertstore==0.2
WTForms==3.0.1
